Add documentation of handlers
The visitor interface was dropped more than two years ago in commits
445cafa3109477023d36bc44cda3e83dea06bb09 and
b4394905b37d963c4908caf0dc8b3c4ec0785544 of libosmium on June 18, 2014.
visitor.hpp as a file name is the only leftover of that change.
Therefore, I decided to remove "Visitor" from the title of this section.